Tickets/dm 46040 (#19)
* WIP

* temp include adapter code in notebooks for use under Times Square

* ignore some notebooks in Times Square
