--- notebooks_tsqr/SAVE_exposurelog.yaml
@@ -1,16 +1,0 @@
-# For use with a Times Square notebook
-title: Logging and Reporting for Narrative Log
-description: Prototype 1
-authors:
-  - name: Steve Pothier
-    slack: Steve Pothier
-tags:
-  - reporting
-  - prototype
-parameters:
-  record_limit:
-    type: integer
-    description: Max number of records to output
-    default: 99
-    minimum: 1
-    maximum: 9999
