+++ python/lsst/ts/logging_and_reporting/version.py
@@ -1,3 +1,3 @@
 # Generated by setuptools_scm
 __all__ = ["__version__"]
-__version__ = "0.1.dev26+g8a2fd11.d20240903"
+__version__ = "0.1.dev28+g968997e.d20240904"
